+++ src/app/(auth)/page.module.scss
