+++ src/app/components/Layouts/Content.module.scss
@@ -4,3 +4,9 @@
   background-color: var(--color-white);
   background-size: 32px 32px;
 }
+
+.main {
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+}
