+++ src/app/components/Layouts/Footer.tsx
@@ -1,8 +1,5 @@
-import styles from '@styles/layouts/Footer.module.scss'
+import styles from './Footer.module.scss';
 
 export default function FooterLayout() {
-  return(
-    <footer className={styles.footer}>
-    </footer>
-  );
+  return <footer className={styles.footer}></footer>;
 }
