+++ src/app/components/Layouts/Header.module.scss
@@ -1,0 +1,17 @@
+.header {
+  align-items: center;
+  width: 100%;
+  height: 96px;
+  background-color: gray;
+}
+
+.link {
+  padding: 0 5px;
+  color: var(--color-white);
+}
+
+.nav ul {
+  display: flex;
+  flex-direction: row;
+  list-style-type: none;
+}
