+++ src/app/components/Layouts/Header.tsx
@@ -1,8 +1,28 @@
-import styles from '@styles/layouts/Header.module.scss'
+import styles from './Header.module.scss';
+import Link from 'next/link';
 
 export default function HeaderLayout() {
-  return(
+  return (
     <header className={styles.header}>
+      <nav className={styles.nav}>
+        <ul>
+          <li>
+            <Link className={styles.link} href="/">
+              ホーム
+            </Link>
+          </li>
+          <li>
+            <Link className={styles.link} href="/sign_up">
+              新規登録
+            </Link>
+          </li>
+          <li>
+            <Link className={styles.link} href="/sign_in">
+              サインイン
+            </Link>
+          </li>
+        </ul>
+      </nav>
     </header>
   );
 }
