+++ src/app/components/common/Button.tsx
@@ -1,4 +1,4 @@
-import styles from '@styles/objects/components/button.module.scss';
+import styles from './Button.module.scss';
 
 type ButtonProps = {
   text: string;
