+++ src/app/page.tsx
@@ -1,5 +1,5 @@
 import Image from 'next/image';
-import styles from '../styles/page.module.scss';
+import styles from './page.module.scss';
 
 export default function Home() {
   return (
