+++ src/styles/globals.scss
@@ -1,1 +1,1 @@
-@use './foundation/base.module';
+@use './base.module';
