--- src/styles/layouts/Header.module.scss
@@ -1,5 +1,0 @@
-.header {
-  width: 100%;
-  height: 96px;
-  background-color: gray;
-}
