--- src/styles/layouts/main.module.scss
@@ -1,5 +1,0 @@
-.main {
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
-}
